Correcciones

+++ src/main/java/com/w2m/spaceships/repositories/SpaceshipRepository.java
@@ -4,6 +4,6 @@ import com.w2m.spaceships.entities.Spaceship;
 import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SpaceshipRepository extends JpaRepository<Spaceship,Long> {
+public interface SpaceshipRepository extends JpaRepository<Spaceship, Integer> {
   List<Spaceship> findByNameContaining(String nombre);
 }
